#1 Rename math.hpp→math_basics.hpp, kspc.hpp→math.hpp
